TimeGraph (3rd tab Animation Manager): correction of labels, graph range and positioning; threshold form now remembers its values when re-opened; index is now set correctly if the animation has been running for a while before user clicks the 3rd tab.
AnimationDrawer:
removed deadlock due to lock lasting too long;
Now MapIndex and currentIndex are consistently set to the correct values at all times;
Now master/slave animations (per index) do not lose track; run, stop, pause, frame back, frame forward, rewind, end are all corrected;
In time mode, the map is now activated on-time, and not too early or too late;
In time mode, now the time-index correctly follows the timestamps;
Slave animation offset and scale now works;
Adding/removing slaves now corrects the administration so that a slave can't have other slaves;

AnimationManager:
The content in all four tabs is now updated when changing the active animation; previously the tabs weren't refreshed, leaving the state belonging to the wrong animation.
Adding/removing animations now is also propagated correctly to the administration kept by the AnimationManager; previously one would still see after-effects after closing an animation window
The speed-control of the animation no longer "blocks" the other tabs when the speed is more than 5.0 fps
The speed-control now remembers the speed when switching between animations;
Consistent behavior when user clicks "Stop" in a slave animation (the slave is un-linked, leaving only the master running)
The animation step (in the 2nd tab offset/step) now works (previously it was ignored)
Now a slow animation can also be used to synchronize a fast one (the fast one will skip frames).
The indexes that are displayed to the user now start at 1, and are the same in all relevant views (1st map corresponds to text "index 1" on the top of the mapwindow and index 1 in the progress tab).
Added option "No Threshold" in the progress tab, to be able to "undo" a threshold when one no longer wants it.


git-svn-id: https://svn.52north.org/svn/ilwis/drawers@3212 b64c5c0c-1e74-440b-9268-4ad09cf26390